+++ engine/src/strategy_engine/node/indicator_node/mod.rs
@@ -32,7 +32,7 @@ pub struct IndicatorNode {
 
 impl IndicatorNode {
     pub fn new(
-        strategy_id: i64, 
+        strategy_id: i32, 
         node_id: String, 
         node_name: String, 
         trade_mode: TradeMode,
