+++ engine/src/strategy_engine/node/position_node/mod.rs
@@ -30,7 +30,7 @@ pub struct PositionNode {
 
 impl PositionNode {
     pub fn new(
-        strategy_id: i64,
+        strategy_id: i32,
         node_id: String,
         node_name: String,
         trade_mode: TradeMode,
