+++ engine/src/strategy_engine/strategy/live_strategy/live_strategy_function/add_indicator_node.rs
@@ -17,7 +17,7 @@ impl LiveStrategyFunction {
     pub async fn add_indicator_node(
         graph: &mut Graph<Box<dyn NodeTrait>, (), Directed>, 
         node_indices: &mut HashMap<String, NodeIndex>,
-        strategy_id: i64,
+        strategy_id: i32,
         node_id: String, 
         node_name: String, 
         live_config: Option<IndicatorNodeLiveConfig>,
